Reorganize tests to match src directory structure
